fix manual API tests
