+++ .gitlab/tests/contrib.yml
@@ -193,11 +193,6 @@ tornado:
   variables:
     SUITE_NAME: "tornado"
 
-unittest:
-  extends: .test_base_riot_snapshot
-  variables:
-    SUITE_NAME: "unittest"
-
 wsgi:
   extends: .test_base_riot_snapshot
   variables:
