+++ tests/ci_visibility/api/test_api_fake_runners.py
@@ -6,6 +6,7 @@ import pytest
 
 from ddtrace.internal.ci_visibility.recorder import _CIVisibilitySettings
 from tests import utils
+from tests.ci_visibility.util import _get_pytest_snapshot_gitlab_ci_env_vars
 from tests.utils import TracerTestCase
 from tests.utils import override_env
 from tests.utils import snapshot
@@ -50,15 +51,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_all_pass.py"])
+            subprocess.run(["python", "fake_runner_all_pass.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_all_fail(self):
@@ -69,8 +73,12 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
         ):
             with mock.patch(
@@ -88,15 +96,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_all_skip.py"])
+            subprocess.run(["python", "fake_runner_all_skip.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_all_itr_skip_test_level(self):
@@ -107,15 +118,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_all_itr_skip_test_level.py"])
+            subprocess.run(["python", "fake_runner_all_itr_skip_test_level.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_all_itr_skip_suite_level(self):
@@ -126,15 +140,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_all_itr_skip_suite_level.py"])
+            subprocess.run(["python", "fake_runner_all_itr_skip_suite_level.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_mix_pass(self):
@@ -145,15 +162,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_mix_pass.py"])
+            subprocess.run(["python", "fake_runner_mix_pass.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_mix_fail(self):
@@ -164,15 +184,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_mix_fail.py"])
+            subprocess.run(["python", "fake_runner_mix_fail.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_mix_fail_itr_test_level(self):
@@ -183,15 +206,18 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_mix_fail_itr_test_level.py"])
+            subprocess.run(["python", "fake_runner_mix_fail_itr_test_level.py"])
 
     @snapshot(ignores=SNAPSHOT_IGNORES)
     def test_manual_api_fake_runner_mix_fail_itr_suite_level(self):
@@ -202,12 +228,15 @@ class FakeApiRunnersSnapshotTestCase(TracerTestCase):
         self.testdir.chdir()
 
         with override_env(
-            dict(
-                DD_API_KEY="foobar.baz",
+            _get_pytest_snapshot_gitlab_ci_env_vars(
+                dict(
+                    DD_API_KEY="foobar.baz",
+                    CI_PROJECT_DIR=str(self.testdir.tmpdir),
+                    DD_CIVISIBILITY_AGENTLESS_ENABLED="false",
+                )
             )
+        ), mock.patch(
+            "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
+            return_value=_CIVisibilitySettings(False, False, False, False),
         ):
-            with mock.patch(
-                "ddtrace.internal.ci_visibility.recorder.CIVisibility._check_settings_api",
-                return_value=_CIVisibilitySettings(False, False, False, False),
-            ):
-                subprocess.run(["python", "fake_runner_mix_fail_itr_suite_level.py"])
+            subprocess.run(["python", "fake_runner_mix_fail_itr_suite_level.py"])
