+++ tests/snapshots/tests.contrib.pytest.test_pytest_snapshot.test_pytest_with_ddtrace_patch_all.json
@@ -10,7 +10,7 @@
     "error": 0,
     "meta": {
       "_dd.p.dm": "-0",
-      "_dd.p.tid": "66e0544900000000",
+      "_dd.p.tid": "66e0558600000000",
       "component": "httpx",
       "http.method": "GET",
       "http.status_code": "404",
@@ -18,7 +18,7 @@
       "http.useragent": "python-httpx/0.27.0",
       "language": "python",
       "out.host": "localhost",
-      "runtime-id": "0512abfa031a4dc98bb30770e36b96b1",
+      "runtime-id": "1b0edee6b1c942e084b30c461749f778",
       "span.kind": "client"
     },
     "metrics": {
@@ -26,8 +26,8 @@
       "_dd.top_level": 1,
       "_dd.tracer_kr": 1.0,
       "_sampling_priority_v1": 1,
-      "process_id": 37996
+      "process_id": 39260
     },
-    "duration": 2947791,
-    "start": 1725977673643562971
+    "duration": 2429750,
+    "start": 1725977990963671590
   }]]
